requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.37.1
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
mysql-connector-python==8.0.25
mysqlclient==2.0.3
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.7.4
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.12
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
